fix : 챌린지 진행률 double값으로 수정

See merge request s07-final/S07P31A106!77

+++ backend/challenge-service/src/main/java/com/ssafy/challengeservice/dto/ChallengeRes.java
@@ -45,7 +45,7 @@ public class ChallengeRes {
         challengeRes.participantsCnt = challenge.getParticipantsCnt();
         challengeRes.rewardPoint = challenge.getRewardPoint();
         challengeRes.progress = challenge.getProgress();
-        challengeRes.progressRate = (challenge.getProgress() / challenge.getGoal()) * 100;
+        challengeRes.progressRate = ((double)challenge.getProgress() / challenge.getGoal()) * 100;
         challengeRes.finishFlag = challenge.getFinishFlag();
         challengeRes.imageUrl = challenge.getImageUrl();
         challengeRes.challengeId = challenge.getId();
